+++ drools-model/drools-model-compiler/src/main/java/org/drools/modelcompiler/builder/generator/DrlxParseUtil.java
@@ -533,92 +533,6 @@ public class DrlxParseUtil {
         return ruleContext.getDelarationType(variableName);
     }
 
-    public static TypedExpression toMethodCallWithClassCheck(RuleContext context, Expression expr, String bindingId, Class<?> clazz, TypeResolver typeResolver) {
-
-        final Deque<ParsedMethod> callStackLeftToRight = new LinkedList<>();
-
-        createExpressionCall(expr, callStackLeftToRight);
-
-        java.lang.reflect.Type previousClass = clazz;
-        Expression previousScope = null;
-
-        for (ParsedMethod e : callStackLeftToRight) {
-            if (e.expression instanceof NameExpr || e.expression instanceof FieldAccessExpr || e.expression instanceof NullSafeFieldAccessExpr) {
-                if (e.fieldToResolve.equals( bindingId )) {
-                    continue;
-                }
-                if (previousClass == null) {
-                    try {
-                        previousClass = typeResolver.resolveType( e.fieldToResolve );
-                        previousScope = new NameExpr( e.fieldToResolve );
-                    } catch (ClassNotFoundException e1) {
-                        // ignore
-                    }
-                    if (previousClass == null) {
-                        previousClass = context.getDeclarationById( e.fieldToResolve )
-                                .map( DeclarationSpec::getDeclarationClass )
-                                .orElseThrow( () -> new RuntimeException( "Unknown field: " + e.fieldToResolve ) );
-                        previousScope = e.expression;
-                    }
-                } else {
-                    TypedExpression te = nameExprToMethodCallExpr( e.fieldToResolve, previousClass, previousScope );
-                    if (te == null) {
-                        context.addCompilationError( new InvalidExpressionErrorResult( "Unknown field " + e.fieldToResolve + " on " + previousClass ) );
-                        return null;
-                    }
-                    java.lang.reflect.Type returnType = te.getType();
-                    previousScope = te.getExpression();
-                    previousClass = returnType;
-                }
-            } else if (e.expression instanceof MethodCallExpr) {
-                java.lang.reflect.Type returnType = returnTypeOfMethodCallExpr(context, typeResolver, (MethodCallExpr) e.expression, previousClass, null);
-                MethodCallExpr cloned = ((MethodCallExpr) e.expression.clone()).setScope(previousScope);
-                previousScope = cloned;
-                previousClass = returnType;
-            }
-        }
-
-        return new TypedExpression(previousScope, previousClass);
-    }
-
-    private static Expression createExpressionCall(Expression expr, Deque<ParsedMethod> expressions) {
-
-        if (expr instanceof NodeWithSimpleName) {
-            NodeWithSimpleName fae = (NodeWithSimpleName)expr;
-            expressions.push(new ParsedMethod(expr, fae.getName().asString()));
-        }
-
-        if (expr instanceof NodeWithOptionalScope) {
-            final NodeWithOptionalScope<?> exprWithScope = (NodeWithOptionalScope) expr;
-            exprWithScope.getScope().ifPresent(expression -> createExpressionCall(expression, expressions));
-        } else if (expr instanceof FieldAccessExpr) {
-            // Cannot recurse over getScope() as FieldAccessExpr doesn't support the NodeWithOptionalScope,
-            // it will support a new interface to traverse among scopes called NodeWithTraversableScope so
-            // we can merge this and the previous branch
-            createExpressionCall(((FieldAccessExpr) expr).getScope(), expressions);
-        }
-
-        return expr;
-    }
-
-    static class ParsedMethod {
-        final Expression expression;
-
-        final String fieldToResolve;
-
-        public ParsedMethod(Expression expression, String fieldToResolve) {
-            this.expression = expression;
-            this.fieldToResolve = fieldToResolve;
-        }
-        @Override
-        public String toString() {
-            return "{" +
-                    "expression=" + expression +
-                    ", fieldToResolve='" + fieldToResolve + '\'' +
-                    '}';
-        }
-    }
-
     public static Type classToReferenceType(Class<?> declClass) {
         String className = declClass.getCanonicalName();
         return classNameToReferenceType(className);
